requirements.txt
PySimpleGUI==4.45.0
pystray
Pillow~=9.1.0
setuptools~=60.10.0
cx_freeze